chore: refactor host components/hooks
