+++ pages/host/queue/index.tsx
@@ -1,70 +1,18 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
 import Image from 'next/image';
 import plebFMLogo from '../../../public/plebfm-logo.svg';
 import bokeh4 from '../../../public/pfm-bokeh-4.jpg';
-import { getSession, signIn, useSession } from 'next-auth/react';
+import { getSession } from 'next-auth/react';
 import WebPlayback from '../../../components/Leaderboard/SpotifyPlayback';
 import { GetServerSidePropsContext } from 'next';
 import Head from 'next/head';
-import { useRouter } from 'next/router';
 import { Song } from '../../../components/Leaderboard/Song';
 import { QR } from '../../../components/Leaderboard/Qr';
 import { Notifications } from '../../../components/Leaderboard/Notifications';
-import { getLeaderboardQueue, SongObject } from '../../../utils/songs';
-import { findHost } from '../../../utils/host';
+import { useLeaderboard } from '../../../components/hooks/useLeaderboard';
 
 export default function Queue() {
-  const { data: session, status } = useSession();
-  const [queueData, setQueueData] = useState<SongObject[]>([]);
-  const [refreshQueue, setRefreshQueue] = useState<boolean>(true);
-  const [host, setHost] = useState<string>();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (!router) return;
-    if (status === 'unauthenticated') {
-      void signIn('spotify');
-    }
-    if (status === 'authenticated') {
-      //@ts-ignore
-      const spotifyId = session?.user?.id;
-      if (!spotifyId) return;
-      findHost(spotifyId).then(host => {
-        if (!host) {
-          console.error('HOST NOT FOUND');
-          router.push('/host?error=host_not_found');
-        } else {
-          setHost(host.shortName);
-        }
-      });
-    }
-  }, [status, session, router]);
-
-  useEffect(() => {
-    console.log('Session', session);
-    if (session?.error === 'RefreshAccessTokenError') {
-      signIn('spotify');
-    }
-    // const foo = async () => {
-    //   console.log(session)
-    //   //@ts-ignore
-    //   const accessToken = session.accessToken ?? '';
-    //   if (!accessToken) console.warn('ACCESS TOKEN MISSING FOR SPOTIFY');
-    // };
-    // foo();
-  }, [session]);
-
-  useEffect(() => {
-    if (!host || !refreshQueue) return;
-    getLeaderboardQueue(host).then(res => {
-      if (res) setQueueData(res);
-    });
-    setRefreshQueue(false);
-  }, [host, refreshQueue]);
-
-  const refresh = useCallback(() => {
-    setRefreshQueue(true);
-  }, []);
+  const { queueData, refresh, session, host } = useLeaderboard();
 
   return (
     <>
@@ -103,7 +51,7 @@ export default function Queue() {
             </div>
 
             <div className="relative z-20">
-              <Notifications refreshQueue={refresh} />
+              <Notifications refreshQueue={refresh} host={host ?? ''} />
             </div>
 
             {session?.accessToken && host && (
