+++ utils/host.ts
@@ -12,3 +12,26 @@ export const findHost = async (spotifyId: string): Promise<Host> => {
   const resJson = await res.json();
   return resJson?.hosts[0];
 };
+
+export const syncJukebox = async (
+  accessToken: string,
+  shortName: string,
+  deviceId: string,
+) => {
+  let url = `/api/leaderboard/queue`;
+  const body = JSON.stringify({
+    shortName: shortName,
+    accessToken: accessToken,
+    deviceId: deviceId,
+  });
+  const response = await fetch(url, {
+    method: 'POST',
+    body: body,
+    headers: {
+      'Content-Type': 'application/json',
+      Accept: '*/*',
+    },
+  });
+  const res = await response.json();
+  return res;
+};
